Fixing file names to JSX, misc
