--- client/src/components/button/button.js
@@ -1,12 +1,0 @@
-import './button.css';
-
-const InputButton = ({text, ...props}) => {
-
-    return (
-        <button className='button' {...props}>
-            {text}
-        </button>
-    );
-};
-
-export default InputButton;
