+++ client/src/components/drop down/drop-down.css
@@ -26,7 +26,7 @@
     -moz-appearance: none;
     appearance: none;
     width: 300px;
-    height: 50px;
+    height: 40px;
     font-size: 16px;
     color: rgb(50,50,50);
     padding: 3px 10px;
