+++ client/src/components/header/header.jsx
@@ -2,7 +2,7 @@ import './header.css';
 import React, { useState } from "react";
 import Link from '../links/link';
 import { Settings, UserAvatar, } from "@carbon/icons-react";
-import TextField from '../text field/textfield';
+import TextField from '../text-field/text-field';
 
 const Header = () => {
