+++ client/src/components/radio button/radio-button.jsx
@@ -1,5 +1,5 @@
 import React, {useState} from "react";
-import './radiobutton.css';
+import './radio-button.css';
 
 const RadioButton = ({options, ...props}) => {
