+++ client/src/components/text area/text-area.css
