+++ client/src/components/text-field/text-field.css
@@ -1,22 +1,47 @@
-.text-field-group {
-  display: flex;
-  flex-direction: column;
+
+* {
+    font-family: 'Poppins', sans-serif;
 }
 
-.text-field-label {
-  margin-left: 12px;
-  font-size: 12pt;
+.input-group {
+    position: relative;
+    margin: 0 !important;
 }
 
-.text-field {
-  border: 2px solid black;
-  border-radius: 15px;
-  outline: none;
-  padding: 8px 10px;
-  font-size: 12pt;
+.input-group label {
+    position: absolute;
+    top: 50%;
+    left: 10px;
+    transform: translateY(-50%);
+    font-size: 16px;
+    color: rgb(50,50,50);
+    padding: 0 5px;
+    pointer-events: none;
+
 }
 
-.text-field:focus {
-  outline: 2px solid #531CB3;
-  border-color: #531CB3;
+.input-group input {
+    width: 300px;
+    height: 40px;
+    font-size: 16px;
+    color: rgb(50,50,50);
+    padding: 3px 10px;
+    background: transparent;
+    border: 1.5px solid rgb(50,50,50);
+    outline: none;
+    border-radius: 15px;
+}
+
+.input-group input:focus~label,
+.input-group input:valid~label {
+    top: 0;
+    font-size: 12px;
+    background: white;
+    transition: 0.5s ease;
+}
+
+.input-group input:focus,
+.input-grtoup input:valid {
+    border: 1.5px solid #531CB3;
+    transition: 0.5s ease;
 }
