+++ client/src/components/text/text.jsx
